+++ backend/services/agent/app/controllers/caseController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { pool } from "../db/pool";
+import { casesToJSON } from "../lib/utils";
 
 //get all cases as an agent
 export const getCases = async (req: Request, res: Response) => {
@@ -11,26 +12,28 @@ export const getCases = async (req: Request, res: Response) => {
 
         if (cases.rows.length === 0 || !cases.rows) {
             res.status(204).send();
+        } else {
+            const casesJSON = casesToJSON(cases);
+            res.status(200).json({ cases: casesJSON });
         }
 
-        const casesJSON: string | any = [];
-        for (let i = 0; i < cases.rows.length; i++) {
-            const case_id = cases.rows[i].case_id;
-            const title = cases.rows[i].title;
-            const description = cases.rows[i].description;
-            const agent_id = cases.rows[i].agent_id;
-            const customer_id = cases.rows[i].customer_id;
+        // const casesJSON: string | any = [];
+        // for (let i = 0; i < cases.rows.length; i++) {
+        //     const case_id = cases.rows[i].case_id;
+        //     const title = cases.rows[i].title;
+        //     const description = cases.rows[i].description;
+        //     const agent_id = cases.rows[i].agent_id;
+        //     const customer_id = cases.rows[i].customer_id;
 
-            casesJSON.push({
-                case_id,
-                title,
-                description,
-                agent_id,
-                customer_id,
-            });
-        }
-
-        res.status(200).json({ cases: casesJSON });
+        //     casesJSON.push({
+        //         case_id,
+        //         title,
+        //         description,
+        //         agent_id,
+        //         customer_id,
+        //     });
+        // }
+        // res.status(200).json({ cases: casesJSON });
     } catch (e) {
         res.status(500).json({ message: "internal err" });
     }
@@ -46,26 +49,28 @@ export const getResolvedCases = async (req: Request, res: Response) => {
 
         if (cases.rows.length === 0 || !cases.rows) {
             res.status(204).send();
+        } else {
+            const casesJSON = casesToJSON(cases);
+            res.status(200).json({ cases: casesJSON });
         }
+        // const casesJSON: string | any = [];
+        // for (let i = 0; i < cases.rows.length; i++) {
+        //     const case_id = cases.rows[i].case_id;
+        //     const title = cases.rows[i].title;
+        //     const description = cases.rows[i].description;
+        //     const agent_id = cases.rows[i].agent_id;
+        //     const customer_id = cases.rows[i].customer_id;
 
-        const casesJSON: string | any = [];
-        for (let i = 0; i < cases.rows.length; i++) {
-            const case_id = cases.rows[i].case_id;
-            const title = cases.rows[i].title;
-            const description = cases.rows[i].description;
-            const agent_id = cases.rows[i].agent_id;
-            const customer_id = cases.rows[i].customer_id;
+        //     casesJSON.push({
+        //         case_id,
+        //         title,
+        //         description,
+        //         agent_id,
+        //         customer_id,
+        //     });
+        // }
 
-            casesJSON.push({
-                case_id,
-                title,
-                description,
-                agent_id,
-                customer_id,
-            });
-        }
-
-        res.status(200).json({ cases: casesJSON });
+        // res.status(200).json({ cases: casesJSON });
     } catch (e) {
         res.status(500).json({ message: "err, couldn't get all cases" });
     }
